+++ CHANGELOG.md
@@ -126,3 +126,4 @@
 - Virtual Threads are being used in Session Inbound Process Queue and Schedulers
 - It should close sessions whenever exceptions occurred
 - Stopped supporting request queue in ZeroProcessor, so the RequestPolicy is unused
+- Renamed Controller to Processor